python2.7/chapter_8/movie_reviews_analizer_app/webmining_server/pages/serializers.py
from pages.models import SearchTerm
from rest_framework import serializers
        
class SearchTermSerializer(serializers.HyperlinkedModelSerializer):
    
    class Meta:
        model = SearchTerm
        fields = ('id', 'term')
